backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
drf-yasg==1.21.15
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
kombu==5.6.2
langchain==1.2.15
langchain-core==1.2.31
langgraph==1.1.6
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.32
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
prompt_toolkit==3.0.52
pydantic==2.13.1
pydantic_core==2.46.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
qdrant-client==1.17.1
redis==7.4.0
requests==2.33.1
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
APScheduler==3.11.2
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
uuid_utils==0.14.1
vine==5.1.0
wcwidth==0.6.0
xxhash==3.6.0
zstandard==0.25.0
